+++ oap-server/server-storage-plugin/storage-elasticsearch-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/elasticsearch/base/MetricsEsDAO.java
@@ -19,11 +19,14 @@
 package org.apache.skywalking.oap.server.storage.plugin.elasticsearch.base;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.skywalking.library.elasticsearch.response.Documents;
 import org.apache.skywalking.library.elasticsearch.response.search.SearchResponse;
 import org.apache.skywalking.oap.server.core.analysis.DownSampling;
 import org.apache.skywalking.oap.server.core.analysis.TimeBucket;
@@ -35,10 +38,7 @@ import org.apache.skywalking.oap.server.core.storage.type.StorageBuilder;
 import org.apache.skywalking.oap.server.library.client.elasticsearch.ElasticSearchClient;
 import org.apache.skywalking.oap.server.library.client.request.InsertRequest;
 import org.apache.skywalking.oap.server.library.client.request.UpdateRequest;
-import org.apache.skywalking.oap.server.storage.plugin.elasticsearch.IndicesMetadataCache;
 import org.joda.time.DateTime;
-
-import static java.util.stream.Collectors.groupingBy;
 
 @Slf4j
 public class MetricsEsDAO extends EsDAO implements IMetricsDAO {
@@ -52,50 +52,48 @@ public class MetricsEsDAO extends EsDAO implements IMetricsDAO {
 
     @Override
     public List<Metrics> multiGet(Model model, List<Metrics> metrics) {
-        Map<String, List<Metrics>> groupIndices
-            = metrics.stream()
-                     .collect(
-                         groupingBy(metric -> {
-                             if (model.isTimeRelativeID()) {
-                                 // Try to use with timestamp index name(write index),
-                                 // if latest cache shows this name doesn't exist,
-                                 // then fail back to template alias name.
-                                 // This should only happen in very rare case, such as this is the time to create new index
-                                 // as a new day comes, and the index cache is  pseudo real time.
-                                 // This case doesn't affect the result, just has lower performance due to using the alias name.
-                                 // Another case is that a removed index showing existing also due to latency,
-                                 // which could cause multiGet fails
-                                 // but this should not happen in the real runtime, TTL timer only removed the oldest indices,
-                                 // which should not have an update/insert.
-                                 String indexName = TimeSeriesUtils.writeIndexName(model, metric.getTimeBucket());
-                                 // Format the name to follow the global physical index naming policy.
-                                 if (!IndicesMetadataCache.INSTANCE.isExisting(
-                                     getClient().formatIndexName(indexName))) {
-                                     indexName = IndexController.INSTANCE.getTableName(model);
-                                 }
-                                 return indexName;
-                             } else {
-                                 // Metadata level metrics, always use alias name, due to the physical index of the records
-                                 // can't be located through timestamp.
-                                 return IndexController.INSTANCE.getTableName(model);
-                             }
-                         })
-                     );
-
-        // The groupIndices mostly include one or two group,
-        // the current day and the T-1 day(if at the edge between days)
+        Map<String, List<Metrics>> groupIndices = new HashMap<>();
         List<Metrics> result = new ArrayList<>(metrics.size());
-        groupIndices.forEach((tableName, metricList) -> {
-            List<String> ids = metricList.stream()
-                                         .map(item -> IndexController.INSTANCE.generateDocId(model, item.id()))
-                                         .collect(Collectors.toList());
-            final SearchResponse response = getClient().ids(tableName, ids);
-            response.getHits().getHits().forEach(hit -> {
-                Metrics source = storageBuilder.storage2Entity(new HashMapConverter.ToEntity(hit.getSource()));
-                result.add(source);
-            });
-        });
 
+        if (model.isTimeRelativeID()) {
+            metrics.forEach(metric -> {
+                // Try to use with timestamp index name(write index),
+                String indexName = TimeSeriesUtils.writeIndexName(model, metric.getTimeBucket());
+                groupIndices.computeIfAbsent(indexName, v -> new ArrayList<>()).add(metric);
+            });
+
+            Map<String, List<String>> indexIdsGroup = new HashMap<>();
+            groupIndices.forEach((tableName, metricList) -> {
+                List<String> ids = metricList.stream()
+                                             .map(item -> IndexController.INSTANCE.generateDocId(model, item.id()))
+                                             .collect(Collectors.toList());
+                indexIdsGroup.put(tableName, ids);
+            });
+            if (!indexIdsGroup.isEmpty()) {
+                final Optional<Documents> response = getClient().ids(indexIdsGroup);
+                response.ifPresent(documents -> documents.forEach(document -> {
+                    Metrics source = storageBuilder.storage2Entity(new HashMapConverter.ToEntity(document.getSource()));
+                    result.add(source);
+                }));
+            }
+        } else {
+            metrics.forEach(metric -> {
+                // Metadata level metrics, always use alias name, due to the physical index of the records
+                // can't be located through timestamp.
+                String indexName = IndexController.INSTANCE.getTableName(model);
+                groupIndices.computeIfAbsent(indexName, v -> new ArrayList<>()).add(metric);
+            });
+            groupIndices.forEach((tableName, metricList) -> {
+                List<String> ids = metricList.stream()
+                                             .map(item -> IndexController.INSTANCE.generateDocId(model, item.id()))
+                                             .collect(Collectors.toList());
+                final SearchResponse response = getClient().searchIDs(tableName, ids);
+                response.getHits().getHits().forEach(hit -> {
+                    Metrics source = storageBuilder.storage2Entity(new HashMapConverter.ToEntity(hit.getSource()));
+                    result.add(source);
+                });
+            });
+        }
         return result;
     }